Proto/Stake/Storage: new Staking_balance storage table
This new version of the Staking_balance snapshotable storage uses
Stake_repr.Full instead of Tez_repr to represent the staking balances
of delegates. This is to give us information about how the stake of
each delegate was splitted (between frozen and delegated, between
delegate and delegators) when snapshot were taken.