Proto/Stake/Migration: support for migrating at any block
This commit restores the possibility to migrate outside of a cycle
boundary. This should not be needed on mainnet nor on test networks
but the possibility of migrating in the middle of the cycle is assumed
in some migration tests and this patch was simpler than modifying
these migration tests.