Proto/AI: activate new staking parameters after sampling
At cycle switch, in Delegate_cycles.cycle_end, among many other
things, the following happen:

- the staking parameters of a delegate are used to compute its
rights (especially to count how many tez are overcostaked and
overdelegated). This is done in
Delegate_sampler.select_new_distribution_at_cycle_end.

- updates to the staking parameters which have been requested several
cycles earlier are applied. This is done in
Delegate_staking_parameters.activate.

Depending on the order of these events, the staking parameters used
are either the ones for the old cycle or the ones for the new
cycle. This commit changes this from using the new cycle parameters to
using the old cycle parameters so we now use for the computation of
rights the parameters which were active when the corresponding
snapshot was taken.

Closes #5807.