Tests/PBT/Michelson: infrastructure for testing cache invariant
This commit adds a PBT test with the aim of testing the cache
invariant: for any value in the IR, round-tripping through bytes
should produce something which has the same size in the sense of
Script_typed_ir_size.

The test is not very useful yet because the property is only tested on
the unit type. The value of this commit lies in the setting of the
architecture to test the property and in particular in the statement
of the property using the Qcheck2 framework.