WASM: Generalize [reveal] to prepare for the universal reveal
Prior to this patch, the [reveal] type declared in the WASM
interpreter used by the WASM PVM was a copy of the one
declared in the protocol, but where the implementation details
of the [Sc_rollup_reveal_hash.t] type were lost.

With thisp patch, we push the logic a step further. This will
make supporting new reveal mechanisms (like DAL) a lot more
straighforward for the WASM PVM.