[scripts/lint.sh]: Fix [--check-gitlab-ci-yml]
The old version of the script had not been running for a while due to
an error in the CI configuration.

Also, the logic of the old script does not handle multiple files with
conditional includes. Indeed, with conditional include you may want to
define the two jobs with the same name with different contents. This
commit changes the linting to only apply within single files.