Kaitai: fork codec internally
This is a possible temporary hack to allow continued work on kaitai
translation effort whilst additional support for poly-project mono-repo
is being improved.

The main idea is to duplicate the code (and build rules and everything)
of codec so the existing version is not disturbed, but the new version
can be made to support kaitai.