CI: bump opam repos to obtain lighter image for end-to-end tests
We rename the [image_template__runtime_e2etest_dependencies] template
to be coherent with the rename in opam repos.

Note: We switch from [timeout --kill-after] to [timeout -k] since the
new [runtime_e2etest] image does not have coreutils, and instead uses
the busybox timeout binary, which does not have the long form of the
flag. For the same reason, switch from [tr --squeeze-repeats] to [tr
-s].

[scripts/ci/merge_coverage.sh]: make BusyBox compatible