Tezt: test_circular_buffer.ml: use Clap instead of Arg
This is necessary because otherwise we get errors when trying to run
e.g. 'tezt --list'. Indeed, the Arg module doesn't know about '--list'.
It used to work, presumably because Tezt used the Arg module and thus
'--list' was consumed by Arg already?

Anyway, thanks to this change, '--buffer-size' now also appears in '--help',
which is a nice benefit.