Lib_micheline: allow primitives starting with _
This commit extends the Micheline lexer to allow primitives starting
with the underscore symbol (`_`). In particular, this allows to
consider the string "_" as a valid primitive. This has no impact on
Michelson because, in each protocol, Michelson uses a fixed set of
primitives (none of which starting with "_") but this is useful for
extensions using "_" as a pattern such as TZT.