Client/Michelson/Stacks: expand macros before parsing stacks
This commit moves the macro-expansion phase of parsing stacks (and
other_contracts and extra_big_maps) from Michelson_v1_stack to
Client_proto_args.

This makes the corresponding Clic parameters behave more similarly to
the ones used for other commands which is needed to better integrate
them with the rest of the client.