Tezt: fix warning from tezt/tests/retro.ml
This fixes the following warning:

In "tezt/tests/retro.ml", test "Nairobi: Test retro compatibility for format of transfer" was declared with 'codec' in its ~uses but did not call Uses.path on it.

Indeed, the test did nothing if the protocol had no predecessor.
Now the test just does not even register for protocols with no
predecessor.