DAL: Better encoding for POST /slot input
`Data_encoding.string` forces to prefix the contents of the string with its
length. But it turns out our HTTP stack already prefixes the input of a request
with its size. In practise, it means the input needs to be prefixed twice,
which is unfortunate.

You can verify with the following script (`post_dal_file.sh') to generate a
valid Data_encoding encoded string.

    #!/usr/bin/bash

    path="${1}"
    target="${2}"

    file_size="$(cat "${path}" | wc -c)"
    slot_size_bin="$(printf "%08x" "${file_size}")"
    slot_contents="$(cat ${path} | xxd -p)"

    echo -n "${slot_size_bin}${slot_contents}" | xxd -p -r > "${target}"

We can use it to generate a (expectedly valid) input for the RPC from an
arbitrary file.

Before the patch,

    $ ./post_dal_file.sh README.md input.bin
    $ curl localhost:10732/slot --data-binary @input.bin \
        -H 'Content-Type: application/octet-stream'
    # fails with an error “Not enough data”

After the patch

    $ ./post_dal_file.sh README.md input.bin
    $ curl localhost:10732/slot --data-binary @input.bin \
        -H 'Content-Type: application/octet-stream'
    {"commitment":"[…]","commitment_proof":"[…]"}

It works because Variable.string allows to not prefix a string by consuming
every remaining bytes of a byte stream of already known size.