Sequencer: Add a worker to publish blueprints
We add a new worker specifically for publishing blueprints. This way, failing
to publish blueprints won’t prevent the sequencer to operate.

In future patches, we will introduce a resilient publication of blueprints,
which monitor the rollup node state to check that a blueprint that has been
forwarded to the rollup node is eventually injected and processed by the
kernel. But in the meantime, we favor the liveness for users.