Tezt/DAL: print all parameters in Dal_common.make
DAL parameters are only valid when some relations between them are
satisfied, these are checked in
src/lib_crypto_dal/parameters_check.ml.

For instance, dividing the default redundancy_factor by some factor F
requires to either divide by F page_size or number_of_shards or to
multiply by F the slot_size.

This commit extends the error message thrown when DAL parameter check
fails to display the value of all the parameters so that it is easier
to choose which one to change to satisfy the relation.