DAL/Node: fix plugin updating
to fetch the committee for the migration block we need to use the new plugin. So
we need to first update the plugin. Note that the new plugin is not used
directly in the `process_block` function. But `Node_context.fetch_committee`
retrieves it and will use the right plugin.