EVM/Tezt: make observer start more robust
I have observed that the test is flaky for two reasons:
- If the time between blocks is too low, the test will be just the
sequencer producing block, with the observer catching 0 blocks. I
believe the issue arises if you run on a weak machine (i.e. arise when
tests are ran in parallel).
- The test ends by a head's check, which is flaky as well. If the
check is done right after a sequencer block production the heads are
inconsistent righfully.