EVM/Tezt: make random generation test more robust
Ok this one is tricky. As far as I understand some tests are flaky
because the estimateGas does not provide enough gas. However, all
estimateGas are run on the same block but do not provide the same
response, as the block number is constant I make the timestamp
constant as well to make the test more predictible. I am not sure
about the fix. In the worst case it does not change the test semantics.