EVM/Kernel: always keep fallback kernel
The cleanup logic is broken, it cleans only when a migration is
finished. However, an upgrade does not necessarily mean a
migration. But, it is not an issue to keep the fallback kernel as it
can be only:
- Read when the fallback is needed, after an error during a migration.
- Written when the kernel upgrades, therefore overwrites any existing value.
Therefore, this commit makes explicit that the fallback kernel is
never cleaned.