CI: split the [build_arm64] jobs
Before, the generated external job had a rule set that combined rules
intended for the scheduled pipeline and rules intended for the
before_merging pipeline. This commits splits the job into one external
file set per pipeline.

Furthermore, it inlines the definition of these jobs to the pipeline
in which they belong.