CIAO: rework scaffolding to have one big ignore
Once the full pipeline is generated, we want to return a list
containing all the jobs. In the mean time, we generate jobs and put
their definition in external files. To stop CIAO from overwriting the
pipelien definition file ([.gitlab/ci/pipelines/PIPELINE.yml]) we
return the empty list.

In this commit, I rework the code so that there will only be one place
to change when we go from returning the empty list to returning the
full list of jobs.