CIAO: simplify dependencies specification for parallel jobs
It turns out that there is no need to explicitly state the name of
each instantiation of a parallel job for the dependency to work -- if
you always want to depend on each instantiation of the parallel job,
which seems like a reasonable restriction for now.

Consider e.g. the following configuration:

```
matrixd_job:
  parallel:
    matrix:
      - FRUIT: ["apple", "orange"]

vectord_job:
  parallel: 3

matrix_collector:
  needs: ["matrixd_job"]
  dependencies: ["matrixd_job"]

vector_collector:
  needs: ["vectord_job"]
  dependencies: ["vectord_job"]
```

Here, the collector jobs will get artifacts from all versions of the
matrix'd respectively the vector'd job (assuming they produce
artifacts). In other words, there is not need to explicitly depend on
[vectord_job 1/3] or [matrixd_job: [apple]].