EVM Node: Fail on missing blueprint only in sequencer mode
If an observer is started after a while, it will have to catch-up with two
chains: the one built by the sequencer and the one verified by the rollup node.
Nothing prevents the observer node to crawl the latter faster than the former,
which means it will get confirmation of blueprint applications before ever
knowing about said blueprints.

This is why we make the “missing blueprint” event a warning rather than an
error for the observer.

This means we need to refactor some tests that were relying on the previous
behavior.