EVM Kernel: Fix reimbursement to the sequencer pool address
The failsafe storage mechanism implemented recently is quite subtle (and
error-prone). During the production of a block, `/evm/world_state` is copied to
`/tmp/evm/world_state` and *every access to the durable storage is prefixed by
`/tmp`*. This is why the configuration is passed over in the various block
validation function. This logic was not used for the sequencer pool address,
which is read in place at `/evm/sequencer_pool_address` (that is,
`/tmp/evm/sequencer_pool_address`).

We fix this bug by reading the sequencer pool address once, before starting the
block production, and passing the argument over to use site.