EVM Node: Fix off-by-one in the stream RPC for blueprints
When an observer starts, it connects to another EVM node to fetch blueprints,
starting from `next_blueprint_number'.

If (1) it was in sync with its EVM node endpoint, and (2) it is just stooped
and restarted, then from the current implementation of the stream RPC, it is
requesting a blueprint from the future, and the RPC fails with

    [
      {
        "kind": "temporary",
        "id": "failure",
        "msg": "Cannot start watching from a level in the future"
      }
    ]

This patch changes the behavior to allow an observer node to wait for the next
level.

It is not clear why the observer node was not failing without this patch...