EVM Node: Reconnect to the EVM endpoint as observer
As of today, the EVM node in observer mode exits gracefully (with exit code 0)
when its connection to the EVM endpoint is lost.  With this patch, we change
this behavior, to make the observer trying to reconnect regularly (picking a
random cooldown between 0 and 2s).

This effectively makes the observer node more recilient to sequencer
disconnects. Most importantly, this means the tx pool of an observer isn’t lost
when a sequencer stops.