EVM Node: Add the minimum viable implementation of a `replay' command
With this patch, we add the foundation of a replay command which, taken a
level, fetches the state before the blueprint was applied, and try to apply it
again. Obviously (/hopefully/), this will lead to the exact same result as for
the first time it was applied. In future patch, we will add options to
manipulate the state, in particular to replay a blueprint with a different
kernel.