DAL/Slot manager: change error produced by get_slot
The `get_slot` function attempts to reconstruct the slot if fetching
it from the store fails. If reconstruction also fails, the error
before this commit was about the failed reconstruction attempt, after
this commit it is about the failed store fetching of the slot, which
is probably more helpful for users.