DAL/RPC: fix handling of the Not_found case
Not_found errors are supposed to produce the 404 HTTP error code,
contrary to all other errors which produce the 500 error code.

This distinction is handled by passing to `Errors.to_option_tzresult`
a value for the `~none` optional argument which discriminates the
`Not_found` case.

Because we want this for all handlers in the RPC_server module for
which a `Not_found` error is possible, the `to_option_tzresult` is
locally shadowed by a version which sets the `~none`
parameter. Unfortunately, `Errors.to_option_tzresult` is directly
called twice. This has the effect of producing 500 error codes instead
of 404 for two RPCs: post_slot and get_slot_shard.