EVM Node: Fix upgrade detection
We had a long standing bug where the EVM node would not correctly record a
successful upgrade.

This can be easily explained by the fact that the request to record a
successfull upgrade was incorrect.  In SQL, `x = NULL' is always false, as
`NULL' is not equal to anything.  In this patch, we replace `= NULL' by `IS
NULL', which is the correct way to check that a value is indeed `NULL'.

Besides, the business logic around upgrade detection was broken.  The node
always tries to record a successful upgrade when it expects one, before
actually checking it is the right time to record.  We fixes this flaw as well.