Rollup node: decrease GC frequency to 1000 blocks
The context GC time on etherlink is now around 6 minutes. Triggering a
GC every 100 blocks means that the rollup node is performing GC
operations around 45% of the time (but in the background). The new value
is meant to reduce this proportion to below 5%.

Fixes https://gitlab.com/tezos/tezos/-/issues/6415