EVM Node: Switch SQLite3 database to use WAL journal mode
SQLite3 implements several approaches to implement atomic transactions and
rollbacks. Write-ahead logs are a recent addition which is pretty suitable for
concurrent settings.

Switching to this mode will allow to get a RPC server process which can run
concurrently to the node without any issue.

We put this switch behind a feature flag that can be activated by modifying
the `config.json' file manually, to add

    {
        ...
        "experimental_features" :
          { "sqlite_journal_mode": "wal" }
    }