EVM Node: Fix support for `Earliest' in `eth_getBlockByNumber'
Prior to this patch, the RPC `eth_getBlockByNumber' does not interpret
correctly its block parameter argument. More precisely, `Earliest' is
misinterpreted as `Latest'.

Besides, we refine the behavior of the proxy mode. Now, `Earliest' is
supported, but defaults to `Latest'. This reflects the fact that the proxy node
does not keep any local state, and simply exposes the very latest one as stored
by its rollup node.