EVM Node: Add a new `rpc' mode
The goal of this mode is to have a read-only endpoint capable of re-using the
data-dir of another EVM node (either observer or sequencer) in order to answer
RPCs without having to wait for write operations like blueprints application.

As of this patch, it remains an experimental features, not to be used in
production.