EVM Node: Support charset=utf-8 in Content-Type
The EVM node does not support letting the user select their charset as part of
the Content-Type definition. It is arguably a problem that we inherit from
our HTTP stack, and a proper fix should be implemented at this level.

In the meantime, we propose this dirty but working patch to support the UTF-8
charset.