EVM Node: Abort when blocks don’t come quickly to the observer
It is a recurring bug that observers sometimes stop receiving blueprints. It
seems to be a publisher issue, because when it happens, all the observers
connected to this publisher (in our examples, the sequencer) are affected.

This will require more investigation, but in the meantime, we propose a simple
fix: we make the observer node aware of the expected time between blocks, and
we make it wait at most this time plus 1s before assuming there is a bug. When
the timeout is reached, the observer node fallbacks to trying to get a new,
fresh stream.