EVM/Kernel: [block::new] and [block::first_block] needs the gas limit
It should be luckily equivalent. When the block is stored we use
the default value of 3 fields: gas_limit, miner and extraData. As the
default value of gas limit is already 2^50 it works fine, but if you
change default value in the crate ethereum, you'll have a different
gas limit in the opcode and in the block.