EVM Node: Assume [Pending] block is equivalent to [Latest] block
The JSON RPC API of Ethereum nodes distinguishes between the latest block (the
most recent one) and the pending block. Ideally, the pending block is a fresh
block created on top of the [Latest] one, with the contents of the mempool (in
our case, Tx pool).

In this patch, we propose to consider that the pending block is the latest one
(except for eth_getTransactionCount, whose behavior is left unchanged). This is
wrong, strictly speaking, but could give a “close enough” behiavor that is
better suited for EVM compatibility than just failing as it is implemented
right now.