CI: remove redundant [opam_repository_commit_hash] consistency check
The opam-repository images are now built with the same lock file and
hash that we check against here, rendering these checks obsolete.

A similar reasoning applies to the coherency check between
[opam_repository_commit_hash] and [full_opam_repository_tag]:

Previously, the opam-repository images would store the [version.sh]
used to produced them. This was used to check the coherency between
the value of [full_opam_repository_tag] used to produce the images in
[tezos/opam-repository] and the value of [full_opam_repository_tag] in
[scripts/version.sh] on [tezos/tezos] CI pipelines.

With [opam-repository] images built in the same [tezos/tezos] CI
pipelines as they are consumed, this coherency check is
redundant. Consequently, we can also stop storing the [version.sh]
file in the [opam-repository images].