DAL/Test: fail if observer does not receive all the expected shards
This commit makes the amplification_without_lost_shards test
stricter. In this test, all the shards of a slot are sent from a slot
producer DAL node to an observer.

Before this commit, the test was successful if the observer did not
receive all the shards but successfully reconstructed them. Thanks to
https://gitlab.com/tezos/tezos/-/merge_requests/13151, we don't expect
this to happen anymore.