[build-debian-packages]: remove redundant [opam_repository_tag] check
There is no longer any need to check that the
[debian-deps-build.Dockerfile] has been built with the appropriate
state of the repo. This is since the [debian-packages-dependencies]
image is always recreated in the CI when the packages are built.

This change is independent of the rest of this MR, and is a result of
moving the job building these images to a child pipeline where they
are run systematically before packages are built using these images.