EVM Node: Rely on a pool of SQLite connections
This is a preliminary work for introducing concurrent accesses to the node’s
store (SQLite db). We expose the concept of connections directly in
[Evm_store], and we modify the implementation of [Evm_store] to take benefit of
the notion of connections pool.

When used in conjuction with the WAL (write-ahead logging) journal log mode, it
should enable having concurrent read-only accesses to the database.