EVM Kernel: Don’t count ticks to create a [BlockInProgress]
This is to ensure that the behaviors of the EVM node and the rollup nodes
remain in sync.

Consider the following scenario:

    Bluperint N.    Contains 0 transaction
    Bluperint N+1.  Contains 1 transaction estimated to take 30B ticks

In the EVM node, each blueprint is executed in isolation. This means that for
both N and N+1, the cost (in ticks) to create the block in progress is paid in
the first run. As a consequence, the execution of transaction of the blueprint
N+1 is flagged as not retriable. The transaction will be marked as failed
because of OutOfTicks in this situation.

In the rollup node, when both blueprints have been submitted in one Tezos
level, things are different. More precisely, a first [kernel_run] will be used
to

    1. create the first block in progress
    2. execute the block
    3. create the second block in progress
    4. execute the block

During 4. the execution of the execution is marked as retryable, and it will
indeed be retried (because enough ticks has been consumed).

Then, in a brand new [kernel_run], the block in progress for N+1 is already
created, and as a consequence, the execution of the first transaction is
started with the full ticks allowed.

This creates a small window for the transaction to fail from the perspective of
the sequencer, but to succeed as far as the rollup nodes are concerned.

By not accounting for ticks for this one-time step, we remove completely this
window of opportunity. This does not pose a security risk wrt. refutability,
because even in the worst case scenario, the number of ticks unaccounted for is
around 2B, and this worst case cannot happen several time during a
[kernel_run].

Co-authored-by: Pierrick Couderc <pierrick.couderc@nomadic-labs.com>