EVM Node: Remove the TX pool from the hot path of block production
Right now, the TX pool is a real bottleneck for the block production. Both
transaction injection and block production go through the TX pool worker, which
treats requests sequentially.

This does not affect the overall throughput, but it makes block production very
unpredictable. This patch fixes this issue, by removing the TX pool from the
hot path of the transaction injection. The heavy lifting (the validation) is
done outside.

Besides, we modify the `Evm_state` module in order to better leverage the WASM
PVM. More precisely, we want to reduce the interferences of the usage of the
PVM by the RPCs to avoid slowing down the block production. We do that by
leveraging the `Lwt_preemptive` module (used by the fast exec of the WASM PVM)
internals: under the hood, it allocates 4 threads by default. We reserve 1 to
the block production, and 3 for the RPCs (basically).