EVM Node: Add a private method to inject transactions to a tx pool
Submitting a transaction to an EVM node (using `eth_sendRawTransaction`)
requires the transaction to be validated. This includes a signature check,
which is (at the moment) performed inside the kernel, that is, inside the WASM
PVM. Unfortunately, the access to the WASM PVM is a scarce resource.

With this patch, we propose to add a new private (i.e., privileged) RPC, which
allows to inject a transaction in the TX pool of a node directly. This RPC
bypasses the signature check.