Proto/tests: update build rules
This change is pretty straightforward: in lib_protocol/test,
anything for 020 is activated for 019, and anything for 019 and
not for 020 is removed for 019.
The exception to this is test_host_operation in
integration/operations, because it requires the implementation of the
HOST operation, which is not in Paris.
Also we update the README, and expose a RPC in the plugin for the
tests (I'm not doing a separate commit for that)