Proto/tests: backport integration (root) tests
Unchanged from alpha, except test_scenario_slashing_stakers.ml, in
which we change the last allowed cycle for the validation erorr on
late denunciation inclusion. This is because this was changed in alpha
in the protocol. Only the return value of the error was changed (see
Outdated_denunciation in validate.ml).