EVM Node: Disable the fallback mechanism of the Fast Execution
The WASM Fast Execution has been designed to fallback to the WASM PVM when
WASMER raises an exception. This was done for ensuring correctness, to not
stuck the rollup because of something specific to WASMER. This is important for
the rollup node, not so much for the EVM Node (and especially the sequencer).
We would rather not publish a blueprint triggering the (slow) execution of WASM
PVM. The only exception is when a blueprint contains delayed transactions. In
such a case, the trap could be triggered by a delayed transaction, and refusing
to apply the blueprint would mean censoring a transaction.