CIAO: rename "marge-bot pipelines" to "final pipelines"
We always runs one final, rebased pipeline before merging an MR to
master (backport MRs are not concerned with this). This is currently
handled by marge-bot. If we switch to merge trains, then GitLab does
this for us when an MR is added to the train.

There are some special rules we want to apply only in final
pipelines. In practice, this only concerns:

 - the trigger job, which mustn't be manual on final pipelines;
 - opam jobs, of which we run more one final pipelines;
 - the coverage report job, which is disabled on final pipelines as
   there is no point in collating a coverage report for an MR that
   will be merged.

Throughout the CI configuration code, we've referred to these final
pipelines as "marge-bot pipelines". To adapt to the potential switch
to merge trains, we're renaming such pipelines to "final pipelines".