EVM Node: Propagate kernel upgrades through blueprints
Currently, EVM observers get their kernel upgrade events via the rollup node,
but these upgrades are applied when blueprints come from their EVM endpoint.
This works well enough when (1) the cooldown period of the upgrade is large
enough, and (2) observers are online and synced when it is injected. This does
not work so well for offline observers, which needs to catchup two streams (EVM
endpoint and rollup node). In that case, nothing ensures the upgrade event will
be fetched before the blueprint activating it to be applied. The EVM node does
not have a way to recover from this easily.

With this patch, we change the source from which EVM nodes in observer mode
fetch the upgrade. The upgrades are propagated through the blueprints
responsible for activating them. It would arguably be better to propagate them
sooner, but the change would be more intrusive and widespread (e.g., changes in
the threshold encryption support modules would be necessary).