[restrict_export_to_octez_source.sh]: anchor paths to root
Anchor the paths in [export-ignore] git attributes written to
[.gitattributes] by [restrict_export_to_octez_source.sh], by prefixing
them with slash. If the paths are not anchored, they match files
throughout the repo. For instance, the line [README.md export-ignore]
does not just remove [README.md] in the root from the export, it
ignores all files named [README.md]. However, the intent of this
script is to only exclude files found at the root (by [find]).