Proto: Update Smart Rollups constants based on a ratio
In P, Smart Rollups constants were updated based on the block time to
match the two weeks commitment period. This has proven to be a mistake,
notably because it becomes hard to test withdrawals on testnets.

As a consequence, we propose with this patch to preserve the commitment
period duration of the previous protocol, instead of enforcing two weeks
everywhere.