Proto: Consider past periods to decide the next commitment level
ParisB has demonstrated the Smart Rollup commitment logic is not resilient to
commitment period constant changes, forcing us to release Octez 20.1 to
activate ParisC. The underlying logic error is well explained in [ParisC
announcement], but in a nutshell, the reason can be summarized as follows: in
ParisB, the new constants is used in place where the previous one should.

[ParisC announcement]: https://research-development.nomadic-labs.com/parisC-announcement.html

This patches addresses the root issue, but keeping track of past commitment
periods (starting from ParisC).