Etherlink: Return the code of the created contract in eth_call
This patch addresses an [issue] reported by partner on how `eth_call` deals
with contract creation. More precisely, on other chains, `eth_call` will return
the bytecode of the created contract, while it won’t on Etherlink (both
Ghostnet and Mainnet).

We add a test to track regression on this question, based on the inputs
provided by the partners in the issue.

[issue]: https://github.com/safe-global/safe-singleton-factory/issues/545

Co-authored-by: Rodi-Can Bozman <rodi.bozman@functori.com>