EVM Kernel: Introduce `StorageVersion`
We encode storage version using a dedicated type, as a preliminary step to
explicitely associate migration function to storage version.

This patch should be a no-op as far as the migration logic is concerned. In
particular, it is expected to be backward compatible with what is currently
running on Mainnet and Ghostnet. However, we use this opportunity to
re-introduce the migration from V11 to V12, which was removed in the previous
kernel freeze. This effectively allows the kernel to be deployed on testnet
still on V11 (which is the case for Mainnet, as it turns out).