EVM Kernel: Apply as many migrations as needed
Initially, we intended to make the very same deployment on both the Testnet and
the Mainnet, and as a consequence, we kept only the latest migration in the
kernel. Since then, reality came and showed us it won’t be that easy. As of
today, it is very likely that the current Testnet version won’t ever be
deployed on Mainnet.

The current migration logic is clearly not ready to handle that, but
fortunately, it does not require heavy changes to fix this, as the size of this
patch demonstrates.

In a nutshell, once a given migration (implemented in `migrate_to`) returns
`MigrationStatus::Done`, we keep updating the stored storage version, before
returning `MigrationStatus::InProgress`. As a result, the caller in `lib.rs`
requests a reboot. This will allow to recall the stage 0 code as part as a
brand new `kernel_run`. Eventually, the stored storage version will catch-up,
and the stage 1 will be able to follow its course.