EVM Node: Fix catchup for observers whose local head is too old
When we have implemented the streamed RPCs for the blueprints, we thought we
were more clever than everyone else and we added an argument to say the
starting point. This makes the logic of the callee way simpler.

Sadly, it turns out we were not that smart, because this approach, albeit more
“elegant” locally, does not scale well when the local head of the EVM node in
observer mode is too old compared to the remote head. In that case, the remote
node sends very quickly a large number of blueprints, and it looks like this
destroys the performance of our client HTTP stack. My uneducated guess: the
stack is trying to parse these blueprints all at once, and it’s terrible at it
(in particular, lots of allocation if we can trust flame graphs).

This patches changes the logic of the observer mode to fallback to the RPC to
fetch blueprints one by one when the remote head is too far in the future.