EVM Node: Add an unsafe command to patch the kernel
The kernel is a key piece of Etherlink, and the harder to update, as it
requires an onchain upgrade. This poses a challenge for the features which are
not related to block validation, but partly implemented with the kernel (this
is typically the case for any RPC requiring to execute EVM code).

To unblock partners and tests these features, we propose to add a command to
patch the kernel used by an EVM node. It is the responsibility of the user of
this command to choose a kernel whose semantics remains consistant with the
on actually deployed on the rollup.