EVM Node: Add a missing case for the `Evm_context` request type
The encoding is only used, to the best of my knowledge, as part of the debug
log of the worker, to advertise that it has finished to treat a request. It is,
as a consequence, never used to _decode_ a value. This is fortunate, because
the `Reconstruct` constructor contains value that cannot be constructed or
mocked easily. As a consequence, we propose a `case` that can only be used to
encode values.