EVM/Tests: fix test that estimateGas enforces limits
Basically the previous test wasn't working properly: it made assumptions on the
tick limit with a recursive call of depth 5000, while it has been limited to 256
recently. Fortunately the sequence of reverts made the contract fail with an
OutOfTicks error while it actually never tested for the gas limit.

The test has been rewritten with an actual test of the limit: the contract
contains a single function that reverts if the gas left is less that 30M.
Calling the simulation without a limits enforces the 30 million limit, thus
makign the contract call fail. Calling with a limit higher than 30M will also
fail, as the node will enforce it at its maximum in that case.