WASM PVM: Check for floats in the Fast Exec runtime
There has been a long standing issue in the rollup node, where it would happily
use WASMER to execute a kernel featuring floats. Under the hood, the culprit is
actually the Fast Exec runtime, which is not defensive enough when loading a
new kernel.

This patch should fix this shortcoming, and ensure no commitment are wrongfully
posted when a rollup ends up with a kernel featuring floats.