WASM PVM: Add a new hook to run when a kernel relies on floats
Following 40721cae3572f5aabbba1de4a37f441528fdeca2, the Fast Execution runtime
of the WASM PVM now checks the kernel complies with the PVM constraints (most
notably, the kernel should not use floats). When it happens, the Fast Execution
runtime is not used, and instead, the WASM PVM is used as a fallback.

This way, the semantics of smart rollups is preserved, _but_ it can be pretty
confusing from a sys admin perspective. To improve the situation, we introduce
a new hook that is called when the typechecking of the WASM PVM fails.

We use this hook in the EVM node to make it explicit when such failure happens.